告警修改

+++ camera/vdi_base/common/adapter/platform/v4l2/src/device_manager/sensor_controller.cpp
@@ -645,7 +645,7 @@ RetCode SensorController::SendAWBLockMetaData(common_metadata_header_t *data)
         sensorVideo_->QuerySetting(GetName(), V4L2_CID_3A_LOCK, reinterpret_cast<int*>(&curLock));
         if (awbLock == OHOS_CAMERA_AWB_LOCK_ON) {
         // set the position of AWB bit to 1;
-            curLock |= static_cast<uint32_t>(V4L2_LOCK_WHITE_BALANCE);
+            curLock |= V4L2_LOCK_WHITE_BALANCE;
         } else if (awbLock == OHOS_CAMERA_AWB_LOCK_OFF) {
             // set the position of AWB bit to 0;
             curLock &= ~V4L2_LOCK_WHITE_BALANCE;
@@ -736,7 +736,7 @@ RetCode SensorController::SendExposureModeMetaData(common_metadata_header_t *dat
             aeLock = *(entry.data.u8);
         }
         if (aeLock == 0) {
-            uint32_t curLock = 0;
+            int curLock = 0;
             auto queryResult = sensorVideo_->QuerySetting(GetName(), V4L2_CID_3A_LOCK, &curLock);
             curLock = exposureMode == OHOS_CAMERA_EXPOSURE_MODE_LOCKED ?
                 curLock | V4L2_LOCK_EXPOSURE : curLock & ~V4L2_LOCK_EXPOSURE;
